Add short intros for candidates, use as meta description.
Fixes #149 - provide a friendly information intro for people who arrive without
much context from Google. Also use the same text as the meta description, so it
displays as a snippet in search results.

Text agreed with Joe and now includes details of the area, which isn't there at
the moment. Also now has a friendly intro for candidates from past elections, which
is blank at the moment.

The text is identical in the intro sentence and in the meta description, I've just
stripped out HTML tags in the meta tag.

We should probably use the same block for both the meta description and the
og:description consistently across the app: this is a bigger change so have
ticketed in #159.