Allow migrations to run on all databases
This is a subtle bug: we're using a data migration in the 'core' app
to apply to the sites framework. Because the migration was only allowed
to run on the logger DB, the migration was getting applied there
but the sites framework was looking at the local DB for the changes.