Add icons on candidate cards showing info available.
Fixes #166.

Uses familiar icons for Facebook, Twitter etc, and text for "leaflets", "statements" and "CV".

Uses FontAwesome, with a custom download via Fontello so that we don't include lots of icons we don't need.

Also includes some rejigging of the links around the card `<section>` tag, because strictly speaking a link around a `<section>` is not valid and Chrome was doing odd things.

Doesn't link directly to a candidate's Twitter, Facebook etc card, because browsers really dislike having links inside other links, and if you link each icon separately, then you end up with a large non-clickable area to the right of the icons, which is exactly where the user's thumb naturally goes (there's probably a fiddly way around this, but I'm not sure it's worth it).

Rose did some paper testing and will report back: we can test in wild and adapt as needed.