Show past hustings, even if no video.
Otherwise we need to do a complicated query to figure out whether the card should exist or not (i.e. if any hustings exist either in the future, or with a past video URL). I'm too tired to figure out how to do it in a DRY way right now :-/