Migrate YNR_ID from CharField to IntegerField
The migration here is a little messy, as Django fails to alter indexes
and constrains of models with an FK to this field.

This means there are 'like' indexes made for FKs using
`varchar_pattern_ops` that can't work with an integer type.

I'm a little amazed that Django fails so hard at this, but then I guess
"change the type of the PK for a model" is a bad idea that doesn't get a
lot of use anyway.

The solution is to manually drop the 'like' indexes. The only other
thing done is to drop the unique constraint from mentions to people.
This should be re-created ideally, but the whole mentions app isn't used
anywhere and the upstream source doesn't exist any more, so I'm not very
worried about it