Lazy import models in views
I think this is a bug in pytest-flakes where it's importing
`__init__.py` files too early in the Django bootstrapping process.

This in turn causes the models imports, via the view, but then the
models fail to register because the app registry isn't set up yet.

To get around this I'm using a lazy import for the models, and will
investigate more with pytest-flakes