Match legacy slugs with only numbers in redirects
This catches an otherwise uncaught URL case, where a slug has only
numbers and an ignored slug is provided.

If only numbers are passed then the postcode view picks it up and
redirects to the election.

If the slug is passed with a `.` then the `election_view` picks it up.

In the case of `2010` and `2015` with a slug, this pattern will redirect
to the new slug.