Lookup PostElection objects using ballot_paper_id
This deals with the case that the ballot has been added in the past but
the election or post object has changed. The post will normally be the
thing to change, especially if the division was assigned a GSS code over
the time the election was active, and EE was updated.

Using the ballot_paper_id as a look up will mean that the ballot always
gets imported with the post and election from EE, but it might cause
orphan posts to exist.