Initial ballot importer
This importer moves away from post and election importers, and towards
using ballot as the main object, adding posts and elections (and
memberships / candidacies) in one process.

The result is, running this one script will set up all the _nomination_
information, leaving `import_people` to add person profile information.

More work to be done on it, including only importing recent or
current ballots, or ballots updated recently (like the Person importer).

We also need to reduce the amount of calls to EE, but that might look
like a bit of a change to the EE API – will spend longer thinking about
how best to do this