Add organisation_type to Post
This is a slightly odd thing to do, but we don't have a concept of
"organisation" in this code base at all.

We also don't expose the organisation in the YNR next API.

To deal with this, I've put a field on Post directly, and pulled it from
the EE API when we pull other data in.

Refs https://github.com/DemocracyClub/aggregator-api/pull/124